+++ src/fs/write.c
@@ -58,7 +58,7 @@ zone_nr new_zone;		/* zone # to be inserted */
 
   /* Is 'position' to be found in the inode itself? */
   if (zone < NR_DZONE_NUM) {
-	rip->i_zone[zone] = new_zone;
+	rip->i_zone[(int) zone] = new_zone;
 	rip->i_modtime = clock_time();
 	return(OK);
   }
@@ -109,7 +109,7 @@ zone_nr new_zone;		/* zone # to be inserted */
   b = (block_nr) *zp << scale;
   bp = get_block(rip->i_dev, b, (new_ind ? NO_READ : NORMAL) );
   if (new_ind) zero_block(bp);
-  bp->b_ind[excess] = new_zone;
+  bp->b_ind[(int) excess] = new_zone;
   rip->i_modtime = clock_time();
   bp->b_dirt = DIRTY;
   put_block(bp, INDIRECT_BLOCK);
