+++ src/lib/stderr.c
@@ -4,5 +4,5 @@ char *s;
   char *p = s;
 
   while(*p != 0) p++;
-  write(2, s, p - s);
+  write(2, s, (int)(p - s));
 }
